Add database management and config functionality
A database is now created upon creation of the server.
The user must have MariaDB installed on their machine and a user
created with the the permissions to add, delete, and modify databases.

All database configurations must be in the config file. A sample
config file - config.yaml is provided.

When a connection request is made, all information contained in the
request is inserted into the database.

When a chunk is uploaded, all information regarding the chunk
is inserted into the database.

The upload directory for a chunk is now pulled from the config file
so that the location can easily be changed.