Add client tests This commit contains a lot of refactoring to accommodate Afero IO mocking. Since we weren't initially employing test-driven development, more updates were needed to isolate code for unit tests.
Hash.go has been added to hash chunks using CRC32 in future, as imohash
is better for large files, whereas CRC32 is better for small
(chunk-like) files. For posterity, if there is a chunk collision, then,
and only then should a SHA1 (or similar) hash checksum be made for the
file.