Reassemble chunks
Files are now reassembled on the server when all file chunks have
been received.

If a file name collision occur, a unique file name is created and
the file chunks will be written to that file.

For example:
- test.mp4 would be renamed to test(1).mp4
- test(1).mp4 would be renamed to test(2).mp4
- test(2).jpg would be renamed to test(3).jpg

etc.

Database files have been refactored to separate querying and database
creation. This will make the transition to using golang-migrate cleaner

Cacheing is implemented using Redis. Frequently accessed data can
now be added to the cache.

A wrapper has been created to handle overlapping methods between
the database and cacheing.