Request missing chunks be resent and fix bug
Endpoint in place for client to tell server that it is finished
sending chunks for a specified file. Server will respond with an array
of chunk IDs.

If the array is empty, the upload completed sucessfully. If not, we
need to implement the client sending the chunks.

Upload cache is now back in use. This may be slower but it eliminates
the race condition that was occuring when a map was used to count
how many chunks were received.